first commit; saving work; refactored cli
