--- yeat/cli.py
@@ -1,188 +1,0 @@
-# -------------------------------------------------------------------------------------------------
-# Copyright (c) 2021, DHS. This file is part of YEAT: http://github.com/bioforensics/yeat
-#
-# This software was prepared for the Department of Homeland Security (DHS) by the Battelle National
-# Biodefense Institute, LLC (BNBI) as part of contract HSHQDC-15-C-00064 to manage and operate the
-# National Biodefense Analysis and Countermeasures Center (NBACC), a Federally Funded Research and
-# Development Center.
-# -------------------------------------------------------------------------------------------------
-
-from .bandage import bandage
-from argparse import Action, ArgumentParser, ArgumentTypeError
-import json
-from pathlib import Path
-from pkg_resources import resource_filename
-from snakemake import snakemake
-import sys
-import yeat
-from yeat.config import AssemblerConfig
-
-
-CONFIG_TEMPLATE = [
-    dict(
-        algorithm="spades",
-        extra_args="--meta",
-    ),
-    dict(
-        algorithm="megahit",
-        extra_args="--min-count 5 --min-contig-len 300",
-    ),
-]
-
-
-class InitAction(Action):
-    def __call__(self, parser, namespace, values, option_string=None):
-        json.dump(CONFIG_TEMPLATE, sys.stdout, indent=4)
-        print()
-        raise SystemExit()
-
-
-def run(
-    fastq1,
-    fastq2,
-    assembly_configs,
-    outdir=".",
-    cores=1,
-    sample="sample",
-    dryrun="dry",
-    downsample=0,
-    coverage=150,
-    seed=None,
-    genomesize=0,
-):
-    snakefile = resource_filename("yeat", "Snakefile")
-    r1 = Path(fastq1).resolve()
-    if not r1.is_file():
-        raise FileNotFoundError(f"No such file: '{r1}'")
-    r2 = Path(fastq2).resolve()
-    if not r2.is_file():
-        raise FileNotFoundError(f"No such file: '{r2}'")
-    assemblers = [config.algorithm for config in assembly_configs]
-    extra_args = {config.algorithm: config.extra_args for config in assembly_configs}
-    config = dict(
-        read1=r1,
-        read2=r2,
-        assemblers=assemblers,
-        extra_args=extra_args,
-        sample=sample,
-        downsample=downsample,
-        coverage=coverage,
-        seed=seed,
-        genomesize=genomesize,
-    )
-    success = snakemake(
-        snakefile,
-        config=config,
-        cores=cores,
-        dryrun=dryrun,
-        printshellcmds=True,
-        workdir=outdir,
-    )
-    if not success:
-        raise RuntimeError("Snakemake Failed")
-
-
-def check_positive(value):
-    try:
-        value = int(value)
-        if value <= 0:
-            raise ArgumentTypeError(f"{value} is not a positive integer")
-    except ValueError:
-        raise ArgumentTypeError(f"{value} is not an integer")
-    return value
-
-
-def get_parser(exit_on_error=True):
-    parser = ArgumentParser(exit_on_error=exit_on_error)
-    parser.add_argument("-v", "--version", action="version", version=f"YEAT v{yeat.__version__}")
-    parser.add_argument(
-        "-o",
-        "--outdir",
-        type=str,
-        metavar="DIR",
-        default=".",
-        help="output directory; default is current working directory",
-    )
-    parser.add_argument(
-        "-t",
-        "--threads",
-        type=int,
-        metavar="T",
-        default=1,
-        help="execute workflow with T threads; by default T=1",
-    )
-    parser.add_argument(
-        "--sample",
-        type=str,
-        metavar="S",
-        default="sample",
-        help="specify a unique sample name S for storing assembly results in the working directory; by default S=sample",
-    )
-    parser.add_argument(
-        "-n",
-        "--dry-run",
-        action="store_true",
-        help="construct workflow DAG and print a summary but do not execute",
-    )
-    parser.add_argument(
-        "-d",
-        "--downsample",
-        type=int,
-        metavar="D",
-        default=0,
-        help="randomly sample D reads from the input rather than assembling the full set; set D=0 to perform auto-downsampling to a desired level of coverage (see --coverage); set D=-1 to disable downsampling; by default D=0",
-    )
-    parser.add_argument(
-        "-c",
-        "--coverage",
-        type=check_positive,
-        metavar="C",
-        default=150,
-        help="target an average depth of coverage Cx when auto-downsampling; by default, C=150",
-    )
-    parser.add_argument(
-        "--seed",
-        type=int,
-        metavar="S",
-        default=None,
-        help="seed for the random number generator used for downsampling; by default the seed is chosen randomly",
-    )
-    parser.add_argument(
-        "-g",
-        "--genome-size",
-        type=int,
-        metavar="G",
-        default=0,
-        help="provide known genome size in base pairs (bp); by default, G=0",
-    )
-    parser.add_argument(
-        "--init",
-        action=InitAction,
-        nargs=0,
-        help="print a template assembly config file to the terminal (stdout) and exit",
-    )
-    parser.add_argument("config", type=str, help="configfile")
-    parser.add_argument("reads", type=str, nargs=2, help="paired-end reads in FASTQ format")
-    return parser
-
-
-def main(args=None):
-    if args is None:
-        args = get_parser().parse_args()
-    assembly_configs = AssemblerConfig.parse_json(open(args.config))
-    run(
-        *args.reads,
-        assembly_configs=assembly_configs,
-        outdir=args.outdir,
-        cores=args.threads,
-        sample=args.sample,
-        dryrun=args.dry_run,
-        downsample=args.downsample,
-        coverage=args.coverage,
-        seed=args.seed,
-        genomesize=args.genome_size,
-    )
-    if not args.dry_run:
-        bandage.run_bandage(
-            assembly_configs=assembly_configs, outdir=args.outdir, cores=args.threads
-        )
