+++ yeat/cli/long.py
@@ -1,0 +1,13 @@
+# -------------------------------------------------------------------------------------------------
+# Copyright (c) 2023, DHS. This file is part of YEAT: http://github.com/bioforensics/yeat
+#
+# This software was prepared for the Department of Homeland Security (DHS) by the Battelle National
+# Biodefense Institute, LLC (BNBI) as part of contract HSHQDC-15-C-00064 to manage and operate the
+# National Biodefense Analysis and Countermeasures Center (NBACC), a Federally Funded Research and
+# Development Center.
+# -------------------------------------------------------------------------------------------------
+
+
+def cli(subparsers):
+    parser_long = subparsers.add_parser("long", help="long-reads")
+    parser_long.add_argument("read", type=str, help="long read in FASTQ format")
