+++ yeat/workflows/short/paired/Snakefile
